requirements.txt
flask==1.1.1
python-dotenv==0.12.0
flask-sqlalchemy==1.4.5
Flask-Migrate==2.7.0
ddtrace==0.44.0
python-slugify==5.0.2